refactor(types): align interface types with BE modules and API response
